pyproject.toml
[project]
name = "ort"
version = "0.0.1"
description = "The open route tracker manager"
authors = [
    {name = "Marco Moenig (info@moenig.it)"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "ruff (>=0.9.6,<0.10.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "geoalchemy2 (>=0.17.0,<0.18.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "gpxpy (>=1.6.2,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "uuid (>=1.30,<2.0)",
    "sqlmodel (==0.0.22)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "loguru (>=0.7.2,<0.8.0)",
    "i18n (>=0.2,<0.3)",
    "shapely (>=2.0.7,<3.0.0)",
    "exifread (>=3.0.0,<4.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "pyproj (>=3.7.1,<4.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "geojson-pydantic (>=1.2.0,<2.0.0)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
start = "app.main:app"    

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"